IntroFragment introduced
