+++ casterio/build.gradle
@@ -25,8 +25,6 @@ def RxVersion = '1.1.0'
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:23.2.1'
-    compile 'com.google.code.gson:gson:2.5'
 
     //Networking
     compile "com.squareup.okhttp3:okhttp:$OkHttpVersion"
@@ -34,4 +32,7 @@ dependencies {
     //Rx
     compile "io.reactivex:rxjava:$RxVersion"
     compile "io.reactivex:rxandroid:$RxVersion"
+    compile 'com.android.support:appcompat-v7:23.2.1'
+    compile 'com.google.code.gson:gson:2.5'
+    compile 'com.android.support:support-v4:23.2.1'
 }
